build: migrate golangci-lint to v2 and pin the version
The CI lint job pinned `version: latest` for golangci-lint-action, which
silently drifted from v1 to v2 once v2 was released. v2 cannot parse the v1
config format, so the lint job was effectively broken (and could not run
locally either).

- Migrate .golangci.yaml to v2 format via `golangci-lint migrate`; this also
  drops linters removed/merged in v2 (exportloopref, gosimple, stylecheck).
- Pin golangci-lint to v2.12.2 and bump golangci-lint-action v6 -> v7 (v6 only
  supports v1) so CI and local runs are byte-for-byte identical.
- Resolve all 11 findings v2 surfaced:
  - fix errors.Is argument order (SA1032, a latent bug in two call sites);
  - drop redundant embedded-field selectors (QF1008);
  - extract a repeated error string into a constant (goconst);
  - log the sequence as uint64 to avoid an int64 conversion (gosec G115);
  - annotate the bounded base58 decode conversion with a justified nolint;
  - resolve whitespace/cuddling (wsl).
- Document the pinned version in CLAUDE.md.

+++ app/turl/service.go
@@ -169,7 +169,7 @@ func (c *commandService) Create(ctx context.Context, long []byte) (*model.TinyUR
 	if err != nil {
 		if errors.Is(err, gorm.ErrDuplicatedKey) {
 			slog.Error(fmt.Sprintf("failed to insert into db: %v, try to get from db", err),
-				slog.Any("long url", long), slog.Int64("seq", int64(seq)))
+				slog.Any("long url", long), slog.Uint64("seq", seq))
 
 			record, err = c.db.GetByLongURL(ctx, long)
 			if err != nil {
